+++ lib/scnnr.rb
@@ -14,7 +14,7 @@ require 'scnnr/coordinate'
 require 'scnnr/coordinate/item'
 require 'scnnr/client'
 require 'scnnr/connection'
-require 'scnnr/routing'
+require 'scnnr/routing_helper'
 require 'scnnr/polling_manager'
 require 'scnnr/response'
 require 'scnnr/version'
