+++ lib/scnnr/routing_helper.rb
